chore: Update author ID in createPost function
Update the author ID in the createPost function of post.controllers.js to use a dynamic value instead of a hardcoded ID. This change ensures that the correct author ID is associated with each post and improves the flexibility of the code.

These changes enhance the maintainability and scalability of the application.