chore: Update author ID in createPost function
Refactor the createPost function in post.controllers.js to use a dynamic value for the author ID instead of a hardcoded ID. This change improves the flexibility of the code and ensures that the correct author ID is associated with each post.

These changes enhance the maintainability and scalability of the application.